requirements.txt
absl-py==1.3.0
accelerate==0.30.1
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.1
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitsandbytes==0.43.1
bleach==5.0.1
blis==0.7.9
bson==0.5.10
cachetools==5.2.0
catalogue==2.0.8
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpathlib==0.16.0
comm==0.2.2
confection==0.1.4
contourpy==1.0.6
contractions==0.1.73
cycler==0.11.0
cymem==2.0.7
Cython==0.29.32
dataclasses-json==0.6.6
datasets==2.7.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distro==1.9.0
en-core-web-lg==3.4.1
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.16.2
filelock==3.8.2
fire==0.4.0
flatbuffers==22.11.23
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2024.5.0
gast==0.4.0
gensim==4.2.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
hdbscan==0.8.33
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib-resources==5.10.0
importlib_metadata==7.1.0
install==1.3.5
ipykernel==6.29.4
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpickle==2.2.0
jsonpointer==2.4
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
karateclub==1.3.3
keras==2.11.0
kiwisolver==1.4.4
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langcodes==3.3.0
langsmith==0.1.59
Levenshtein==0.20.8
libclang==14.0.6
lightning-utilities==0.3.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.21.2
matplotlib==3.6.2
matplotlib-inline==0.1.7
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.2
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.6.0
networkx==2.6.3
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.30.0
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.4
pathy==0.6.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==1.4.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGSP==0.5.1
pynndescent==0.5.8
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.20.8
python-louvain==0.16
pytorch-lightning==1.8.2
pytz==2022.6
pyvis==0.3.1
PyYAML==6.0
pyzmq==26.0.3
qtconsole==5.4.0
QtPy==2.2.1
rapidfuzz==2.13.2
rdkit==2022.9.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
sniffio==1.3.1
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.3
SQLAlchemy==2.0.30
srsly==2.4.5
stack-data==0.6.3
sympy==1.12
tenacity==8.3.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-gpu==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.17.0
textsearch==0.0.24
thinc==8.2.3
threadpoolctl==3.1.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.19.1
torch==1.12.0+cu113
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
torch_geometric==2.5.2
torchaudio==0.12.0+cu113
torchmetrics==0.11.4
torchvision==0.13.0+cu113
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.0
triton==2.3.0
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.12.0
tzdata==2024.1
umap-learn==0.5.3
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.19.0
